requirements.txt
fastapi[standard]
openai
python-dotenv
requests